Update 14 files
- /src/test/java/ch/bbbaden/player/PlayerTest.java
- /src/test/java/ch/bbbaden/jeton/JetonTest.java
- /src/test/java/ch/bbbaden/jeton/JetonTypesTest.java
- /src/test/java/ch/bbbaden/casino/CasinoGuiTest.java
- /src/test/java/ch/bbbaden/bingo/BingoGameTest.java
- /src/test/java/ch/bbbaden/bingo/BingoGuiTest.java
- /src/main/java/ch/bbbaden/player/ActualPlayer.java
- /src/main/java/ch/bbbaden/player/ComputerPlayer.java
- /src/main/java/ch/bbbaden/player/Player.java
- /src/main/java/ch/bbbaden/player/PlayerGui.java
- /src/main/java/ch/bbbaden/jeton/Jeton.java
- /src/main/java/ch/bbbaden/jeton/JetonTypes.java
- /src/main/java/ch/bbbaden/casino/CasinoGui.java
- /src/main/java/ch/bbbaden/bingo/BingoGame.java

+++ src/main/java/ch/bbbaden/player/PlayerGui.java
@@ -6,11 +6,11 @@ public class PlayerGui {
 
     private static final Scanner SCANNER = new Scanner(System.in);
 
-    public static double getPlayerStake() {
+    public static int getPlayerStake() {
         while (true) {
             System.out.println("Bitte gib deinen Einsatz ein:");
             try {
-                return SCANNER.nextDouble();
+                return SCANNER.nextInt();
             } catch (RuntimeException e) {
                 System.out.println("Ungültige Eingabe! Bitte gib eine Zahl ein!");
             } finally {
